Updated Get AccessLevel Functions
Added a get AccessLevel by ID function as well as a parameter on
the Get-S2AccessLevels that returns the ID instead of the Names.
I had to run the results of the IDs though a foreach loop of the
Get-S2AccessLevel function to associate ID with Names.

Get-S2AccessLevels -WantKey |
% {
Write-Output "ACCESSLEVELID          : $_";
Get-S2AccessLevel -AccessLevelKey $_
}